refactor survey actions into meta/items
