+++ src/actions/survey/index.test.js
@@ -6,13 +6,11 @@ import {
   SURVEY_LOAD_SUCCESS,
   SURVEY_LOADING,
   SURVEY_CLEAR,
-  META_UPDATE,
   loadSurvey,
   loadingSurvey,
   loadSurveyFailure,
   loadSurveySuccess,
-  clearSurvey,
-  updateMeta
+  clearSurvey
 } from "actions/survey";
 
 describe("actions/survey", () => {
@@ -99,14 +97,6 @@ describe("actions/survey", () => {
   describe("clearSurvey", () => {
     it("dispatches the correct action", () => {
       expect(clearSurvey()).toEqual(actionMatching(SURVEY_CLEAR));
-    });
-  });
-
-  describe("updateMeta", () => {
-    it("should pass 'key' and 'value' payload", () => {
-      const payload = { key: "foo", value: "bar" };
-      const result = updateMeta(payload.key, payload.value);
-      expect(result).toEqual(actionMatching(META_UPDATE, payload));
     });
   });
 });
