+++ src/actions/survey/meta.js
@@ -1,0 +1,11 @@
+export const META_UPDATE = "META_UPDATE";
+
+export function updateMeta(key, value) {
+  return {
+    type: META_UPDATE,
+    payload: {
+      key,
+      value
+    }
+  };
+}
